algorithm updates for python (mostly)
This is a rather larger update for the python version

First update:
if any of the masks are larger than 18% of the entire image then its probably wrong and will then skip that j iteration and go to the next segmentation step.

Second update:
this is when no (or just one) worm is detected from the segmentation. It then lowers the minimun_worm_size from 3000 to 1000 and then tries again