algorithm updates for python (mostly)
This is a rather larger update for the python version

First update:
if any of the masks are larger than 18% of the entire image then its probably wrong and will then skip that j iteration and go to the next segmentation step.

Second update:
this is when no (or just one) worm is detected from the segmentation. It then lowers the minimun_worm_size from 3000 to 1000 and then tries again


Former-commit-id: e3328a826ecec7b2ffc83413334daa757f3049f5 [formerly 6daa20627db504163c7feb1834207bdcd3da67c9]
Former-commit-id: c9d9ec16debb216b190877c7d3213a5ce702fcda
Former-commit-id: 6187df1a1726ac77660221f755fb22d1209025fd